Update avgScore type to double in Stats entity and related serialization; enhance add_stat_view with additional fields for clean sheets, yellow cards, and red cards.

+++ lib/domain/entities/stats.dart
@@ -17,7 +17,7 @@ class Stats {
   int cleanSheets;
   int yellowCards;
   int redCards;
-  int avgScore;
+  double avgScore;
 
   final season = IsarLink<Season>();
   final tournament = IsarLink<Tournament>();
